# docs/swagger.yaml
basePath: /
definitions:
...
  stock.Stock:
    properties:
      id:
        type: string
      item_code:
        type: string
      item_description:
        type: string
      item_id:
        type: string
      quantity:
        type: integer
    type: object
  stock.UpdateStockQuantityRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
  users.ChangePasswordRequest:
    properties:
      id:
        type: string
      password:
        type: string
    required:
    - id
    - password
    type: object
  users.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  users.UpdateUserRequest:
    properties:
      is_active:
        type: boolean
      username:
        type: string
    required:
    - username
    type: object
...
  users.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8125
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API per gestionar compres, ventes i stocks bàsica
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRDY API
  version: "1.0"
paths:
...
  /api/stock:
    get:
      consumes:
      - application/json
      description: Retrieve all stock information (Protected route)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/stock.Stock'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all stocks
      tags:
      - stock
  /api/stock/{item_id}:
    get:
      consumes:
      - application/json
      description: Retrieve stock information for a specific item by its ID (Protected
        route)
      parameters:
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/stock.Stock'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get stock by item ID
      tags:
      - stock
    put:
      consumes:
      - application/json
      description: Update the quantity of stock for a specific item (Protected route)
      parameters:
      - description: Item ID
        in: path
        name: item_id
        required: true
        type: string
      - description: Stock quantity update data
        in: body
        name: quantity
        required: true
        schema:
          $ref: '#/definitions/stock.UpdateStockQuantityRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update stock quantity
      tags:
      - stock
  /api/users:
    get:
      consumes:
      - application/json
      description: Retrieve all users (Protected route)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID (Protected route)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by their ID (Protected route)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information by ID (Protected route)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /api/users/change-password:
    post:
      consumes:
      - application/json
      description: Change the password for a user (Protected route)
      parameters:
      - description: Password change data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/users.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - users
  /api/users/username/{username}:
    get:
      consumes:
      - application/json
      description: Retrieve a user by their username (Protected route)
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by username
      tags:
      - users
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided information (Public route)
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"